Update 8-1-2024
Updated weapon pickups. I created a Gun Checker which will check to see if the player already possesses a weapon or not. If the player is picking up the weapon for the first time, the announcer will not say anything - only when the weapon is selected. However, if the player already has the weapon, if they were to walk over a gun for its ammo, then the announcer will tell the player they had picked up that particular weapon. It's not perfect, however, if you'd like to make some tweaks to it, by all means.